Split contact creation models from view models
