+++ frontend/src/interfaces/IContact.ts
@@ -1,5 +1,3 @@
-import { AddressTypes } from 'constants/addressTypes';
-
 import ITypeCode from './ITypeCode';
 
 export interface IContactCountry {
@@ -62,78 +60,3 @@ export interface IContact {
   person?: IContactPerson;
   organization?: IContactOrganization;
 }
-
-export interface IContactPersonCreate {
-  id?: number;
-  surname: string;
-  firstName: string;
-  middleNames?: string;
-  preferredName?: string;
-  comment?: string;
-  isDisabled: boolean;
-  organizationId?: string;
-  addresses: IContactAddressCreate[];
-  contactMethods: IContactMethodCreate[];
-}
-
-export interface IContactAddressCreate {
-  id?: number;
-  rowVersion?: number;
-  addressType: string;
-  streetAddress1: string;
-  streetAddress2?: string;
-  streetAddress3?: string;
-  regionId?: number;
-  districtId?: number;
-  municipality?: string;
-  provinceId?: number;
-  countryId?: number;
-  countryOther?: string;
-  postal?: string;
-}
-
-export interface IContactMethodCreate {
-  id?: number;
-  rowVersion?: number;
-  contactMethodType: string;
-  value: string;
-}
-
-export const defaultCreatePerson: IContactPersonCreate = {
-  isDisabled: false,
-  firstName: '',
-  middleNames: '',
-  surname: '',
-  preferredName: '',
-  comment: '',
-  contactMethods: [],
-  addresses: [
-    {
-      addressType: AddressTypes.Mailing,
-      streetAddress1: '',
-      streetAddress2: '',
-      streetAddress3: '',
-      municipality: '',
-      countryOther: '',
-      postal: '',
-    },
-    {
-      addressType: AddressTypes.Mailing,
-      streetAddress1: '',
-      streetAddress2: '',
-      streetAddress3: '',
-      municipality: '',
-      countryOther: '',
-      postal: '',
-    },
-    {
-      addressType: AddressTypes.Mailing,
-      streetAddress1: '',
-      streetAddress2: '',
-      streetAddress3: '',
-      municipality: '',
-      countryOther: '',
-      postal: '',
-    },
-  ],
-};
